requirements.txt
# BuildFlow – Dependências principais do backend
# fastapi, uvicorn, SQLAlchemy async, asyncpg, pydantic v2, dotenv, loguru, httpx

fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
asyncpg>=0.29
pydantic>=2.8
python-dotenv>=1.0
loguru>=0.7
httpx>=0.27